pyproject.toml
[tool.poetry]
name = "agentic-platform"
version = "0.1.0"
description = "Multi-tenant AI agentic platform with RAG and MCP support"
authors = ["Platform Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.1.2"
python-multipart = "^0.0.6"
anthropic = "^0.18.0"
openai = "^1.10.0"
pgvector = "^0.2.4"
httpx = "^0.26.0"
tiktoken = "^0.5.2"
pypdf = "^4.0.0"
python-docx = "^1.1.0"
python-magic = "^0.4.27"
aiofiles = "^23.2.1"
beautifulsoup4 = "^4.12.0"
markdown = "^3.5.0"
sentence-transformers = "^2.3.0"
torch = {version = "^2.0.0", optional = true}
python-slugify = "^8.0.0"
aiohttp = "^3.9.0"
structlog = "^24.1.0"
python-json-logger = "^2.0.7"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-httpx = "^0.43b0"
opentelemetry-instrumentation-sqlalchemy = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
sentry-sdk = {extras = ["fastapi", "sqlalchemy", "httpx", "redis"], version = "^1.40.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
factory-boy = "^3.3.0"
locust = "^2.20.0"
bandit = "^1.7.5"
safety = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
